Move list-databases to typescript, to validate that the typescript compiler works
